requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
idna==3.3
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tabulate==0.8.9
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
